+++ packages/core/admin/admin/src/pages/AuthPage/components/AuthButton/index.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+import { Button } from '@strapi/parts/Button';
+
+const AuthButton = styled(Button)`
+  display: inline-block;
+  width: 100%;
+`;
+
+export default AuthButton;
