+++ packages/core/admin/admin/src/pages/AuthPage/components/FieldActionWrapper/index.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+import { FieldAction } from '@strapi/parts/Field';
+
+const FieldActionWrapper = styled(FieldAction)`
+  svg {
+    height: 1rem;
+    width: 1rem;
+    path {
+      fill: ${({ theme }) => theme.colors.neutral600};
+    }
+  }
+`;
+
+export default FieldActionWrapper;
