+++ packages/core/admin/package.json
@@ -41,8 +41,8 @@
     "@strapi/babel-plugin-switch-ee-ce": "1.0.0",
     "@strapi/helper-plugin": "3.6.6",
     "@strapi/utils": "3.6.6",
-    "@strapi/icons": "0.0.1-alpha.10",
-    "@strapi/parts": "0.0.1-alpha.10",
+    "@strapi/icons": "0.0.1-alpha.11",
+    "@strapi/parts": "0.0.1-alpha.11",
     "axios": "^0.21.1",
     "babel-loader": "8.2.2",
     "babel-plugin-styled-components": "1.12.0",
