+++ packages/core/helper-plugin/package.json
@@ -73,6 +73,7 @@
     "babel-plugin-styled-components": "1.12.0",
     "bootstrap": "^4.6.0",
     "classnames": "^2.3.1",
+    "formik": "2.2.9",
     "immutable": "^3.8.2",
     "invariant": "^2.2.1",
     "lodash": "4.17.21",
